+++ demos/manual/src/main/java/is/codion/manual/store/minimal/ui/StoreDemo.java
@@ -85,7 +85,7 @@ public class StoreDemo {
 		UIManager.setLookAndFeel(new MaterialDarker());
 
 		Database database = H2DatabaseFactory
-						.createDatabase("jdbc:h2:mem:h2db",
+						.create("jdbc:h2:mem:h2db",
 										"src/main/sql/create_schema_minimal.sql");
 
 		EntityConnectionProvider connectionProvider =